refactor(abi-model): split lib.rs into per-concern modules
The single 720-line lib.rs becomes the crate doc plus private modules
grouped by ARC-56 concern, with every type re-exported so the public API
stays flat (`algonaut_abi_model::AbiContract`, etc.):

- contract - AbiContract/AbiInterface/network info + genesis_to_network
- method   - AbiMethod (+ get_signature), args, returns, defaults, recommendations
- structs  - named-struct field definitions
- state    - declared global/local/box state
- actions  - OnComplete actions (shared by contract + method)
- events   - ARC-28 events
- program  - deploy/compile metadata (source, bytecode, template/scratch vars)

Tests move next to the types they cover. Pure code movement -- no
behaviour change; the model's tests, the round-trip + schema-validation
suites, and the dependent crates (algonaut_abi, algonaut_abi_macros) are
unchanged and green.
